+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 1.3.0
+
++ Added "Show latest build result" in VSCode tab.
+* Fixed inproper naming in commands
+
 # Version 1.1.2
 
 * Security release